+++ projects/stage-1/middleware-frameworks/my-fault-tolerance/pom.xml
@@ -10,28 +10,22 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>my-interceptor</artifactId>
-    <name>Middleware Frameworks - My Interceptor Framework</name>
+    <artifactId>my-fault-tolerance</artifactId>
+    <name>Middleware Frameworks - My Fault Tolerance Framework</name>
 
     <dependencies>
 
         <!-- Internal Artifacts -->
         <dependency>
             <groupId>org.geekbang.projects</groupId>
-            <artifactId>my-commons</artifactId>
+            <artifactId>my-interceptor</artifactId>
             <version>${revision}</version>
         </dependency>
 
-        <!-- Interceptor API -->
+        <!-- MicroProfile Fault Tolerance -->
         <dependency>
-            <groupId>javax.interceptor</groupId>
-            <artifactId>javax.interceptor-api</artifactId>
-        </dependency>
-
-        <!-- CGLIB -->
-        <dependency>
-            <groupId>cglib</groupId>
-            <artifactId>cglib</artifactId>
+            <groupId>org.eclipse.microprofile.fault-tolerance</groupId>
+            <artifactId>microprofile-fault-tolerance-api</artifactId>
         </dependency>
 
         <!-- Testing -->
@@ -42,5 +36,4 @@
         </dependency>
 
     </dependencies>
-
 </project>
