+++ projects/stage-1/middleware-frameworks/my-fault-tolerance/src/main/resources/META-INF/services/org.geektimes.interceptor.AnnotatedInterceptor
@@ -1,0 +1,6 @@
+org.geektimes.microprofile.faulttolerance.AsynchronousInterceptor
+org.geektimes.microprofile.faulttolerance.BulkheadInterceptor
+org.geektimes.microprofile.faulttolerance.CircuitBreakerInterceptor
+org.geektimes.microprofile.faulttolerance.FallbackInterceptor
+org.geektimes.microprofile.faulttolerance.RetryInterceptor
+org.geektimes.microprofile.faulttolerance.TimeoutInterceptor
