+++ projects/stage-1/middleware-frameworks/my-fault-tolerance/src/test/java/org/geektimes/microprofile/faulttolerance/AsynchronousInterceptorTest.java
@@ -14,41 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geektimes.interceptor.cglib;
+package org.geektimes.microprofile.faulttolerance;
 
-import net.sf.cglib.proxy.MethodProxy;
 import org.geektimes.interceptor.ReflectiveMethodInvocationContext;
+import org.junit.Test;
 
-import javax.interceptor.InvocationContext;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.util.Objects.requireNonNull;
+import java.util.concurrent.Future;
 
 /**
- * {@link InvocationContext} on method using CGLIB
+ * {@link AsynchronousInterceptor} Test
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
  */
+public class AsynchronousInterceptorTest {
 
-class CglibMethodInvocationContext extends ReflectiveMethodInvocationContext {
+    private AsynchronousInterceptor interceptor = new AsynchronousInterceptor();
 
-    private final MethodProxy proxy;
-
-    public CglibMethodInvocationContext(Object target, Method method, MethodProxy proxy, Object... parameters) {
-        super(target, method, parameters);
-        this.proxy = proxy;
-    }
-
-    @Override
-    public Object proceed() throws Exception {
-        try {
-            return proxy.invokeSuper(getTarget(), getParameters());
-        } catch (Throwable throwable) {
-            throw new Exception(throwable);
-        }
+    @Test
+    public void testFuture() throws Throwable {
+        EchoService echoService = new EchoService();
+        Method method = EchoService.class.getMethod("echo", Object.class);
+        ReflectiveMethodInvocationContext context = new ReflectiveMethodInvocationContext
+                (echoService, method, "Hello,World");
+        Future<?> future = (Future) interceptor.execute(context);
+        future.get();
     }
 }
