+++ projects/stage-1/middleware-frameworks/my-fault-tolerance/src/test/java/org/geektimes/microprofile/faulttolerance/BuzServiceImpl.java
@@ -14,41 +14,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geektimes.interceptor.cglib;
-
-import net.sf.cglib.proxy.MethodProxy;
-import org.geektimes.interceptor.ReflectiveMethodInvocationContext;
-
-import javax.interceptor.InvocationContext;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.util.Objects.requireNonNull;
+package org.geektimes.microprofile.faulttolerance;
 
 /**
- * {@link InvocationContext} on method using CGLIB
+ * {@link BuzService} Implementation
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
  */
-
-class CglibMethodInvocationContext extends ReflectiveMethodInvocationContext {
-
-    private final MethodProxy proxy;
-
-    public CglibMethodInvocationContext(Object target, Method method, MethodProxy proxy, Object... parameters) {
-        super(target, method, parameters);
-        this.proxy = proxy;
-    }
+public class BuzServiceImpl implements BuzService {
 
     @Override
-    public Object proceed() throws Exception {
+    public void task() {
         try {
-            return proxy.invokeSuper(getTarget(), getParameters());
-        } catch (Throwable throwable) {
-            throw new Exception(throwable);
+            Thread.sleep(100L);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
         }
+    }
+
+    public void fallbackTask() {
+        System.out.println("fallbackTask");
     }
 }
