+++ projects/stage-1/middleware-frameworks/my-interceptor/src/main/java/org/geektimes/interceptor/jdk/DynamicProxyInterceptorEnhancer.java
@@ -14,26 +14,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geektimes.interceptor;
+package org.geektimes.interceptor.jdk;
 
-import org.junit.Test;
+import org.geektimes.interceptor.InterceptorEnhancer;
 
-import static org.geektimes.interceptor.AnnotatedInterceptor.loadInterceptors;
+import java.lang.reflect.Proxy;
+
+import static java.lang.reflect.Proxy.newProxyInstance;
+import static org.geektimes.commons.reflect.util.ClassUtils.getClassLoader;
 
 /**
- * {@link DefaultInterceptorEnhancer} Test
+ * {@link InterceptorEnhancer} based on JDK Dynamic Proxy
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
  */
-public class DefaultInterceptorEnhancerTest {
+public class DynamicProxyInterceptorEnhancer implements InterceptorEnhancer {
 
-    private InterceptorEnhancer interceptorEnhancer = new DefaultInterceptorEnhancer();
-
-    @Test
-    public void testInterface() {
-        EchoService echoService = new EchoService();
-        echoService = interceptorEnhancer.enhance(echoService, loadInterceptors());
-        echoService.echo("Hello,World");
+    @Override
+    public <T> T enhance(T source, Class<? super T> type, Object... interceptors) {
+        ClassLoader classLoader = getClassLoader(type);
+        return (T) newProxyInstance(classLoader, new Class[]{type}, new InvocationHandlerAdapter(source, interceptors));
     }
 }
