+++ projects/stage-1/middleware-frameworks/my-interceptor/src/main/java/org/geektimes/interceptor/jdk/InvocationHandlerAdapter.java
@@ -30,7 +30,7 @@ import java.lang.reflect.Method;
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since 1.0.0
  */
-public class InvocationHandlerAdapter implements InvocationHandler {
+class InvocationHandlerAdapter implements InvocationHandler {
 
     private final Object source;
