+++ projects/stage-1/middleware-frameworks/my-interceptor/src/test/java/org/geektimes/interceptor/LoggingInterceptor.java
@@ -14,41 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geektimes.interceptor.cglib;
+package org.geektimes.interceptor;
 
-import net.sf.cglib.proxy.MethodProxy;
-import org.geektimes.interceptor.ReflectiveMethodInvocationContext;
-
+import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.util.Objects.requireNonNull;
+import java.util.logging.Logger;
 
 /**
- * {@link InvocationContext} on method using CGLIB
- *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
- * @since 1.0.0
+ * @since
  */
-
-class CglibMethodInvocationContext extends ReflectiveMethodInvocationContext {
-
-    private final MethodProxy proxy;
-
-    public CglibMethodInvocationContext(Object target, Method method, MethodProxy proxy, Object... parameters) {
-        super(target, method, parameters);
-        this.proxy = proxy;
-    }
+@Logging
+@Interceptor
+public class LoggingInterceptor extends AnnotatedInterceptor<Logging> {
 
     @Override
-    public Object proceed() throws Exception {
-        try {
-            return proxy.invokeSuper(getTarget(), getParameters());
-        } catch (Throwable throwable) {
-            throw new Exception(throwable);
-        }
+    protected Object execute(InvocationContext context, Logging logging) throws Throwable {
+        Logger logger = Logger.getLogger(logging.name());
+        logger.info((String) context.getParameters()[0]);
+        return context.proceed();
     }
 }
