--- projects/stage-1/middleware-frameworks/my-interceptor/src/test/java/org/geektimes/interceptor/microprofile/faulttolerance/EchoServiceCglibTest.java
@@ -1,52 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.geektimes.interceptor.microprofile.faulttolerance;
-
-import net.sf.cglib.proxy.Enhancer;
-import net.sf.cglib.proxy.MethodInterceptor;
-import net.sf.cglib.proxy.MethodProxy;
-import org.junit.Test;
-
-import java.lang.reflect.Method;
-
-/**
- * {@link EchoService} CGLIB Test
- *
- * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
- * @since 1.0.0
- */
-public class EchoServiceCglibTest {
-
-    @Test
-    public void test() {
-        Enhancer enhancer = new Enhancer();
-        enhancer.setSuperclass(EchoService.class);
-        enhancer.setCallback(new MethodInterceptor() {
-
-            // @Bulkhead
-
-            @Override
-            public Object intercept(Object obj, Method method, Object[] args, MethodProxy proxy) throws Throwable {
-                return proxy.invokeSuper(obj, args);
-            }
-        });
-
-        EchoService echoService = (EchoService) enhancer.create();
-
-        echoService.echo("Hello,World");
-    }
-}
