+++ projects/stage-1/middleware-frameworks/my-interceptor/src/test/resources/META-INF/services/org.geektimes.interceptor.AnnotatedInterceptor
@@ -1,0 +1,1 @@
+org.geektimes.interceptor.LoggingInterceptor
