+++ projects/stage-1/pom.xml
@@ -43,7 +43,7 @@
         <microprofile-config-api.version>2.0</microprofile-config-api.version>
         <microprofile-rest-client-api.version>2.0</microprofile-rest-client-api.version>
         <microprofile-fault-tolerance-api.version>3.0</microprofile-fault-tolerance-api.version>
-        <microprofile-reactive-messaging-api.version>2.0</microprofile-reactive-messaging-api.version>
+        <microprofile-reactive-messaging-api.version>1.0</microprofile-reactive-messaging-api.version>
 
         <!-- 第三方开源依赖版本-->
         <commons-lang.version>2.6</commons-lang.version>
