+++ projects/stage-1/shopizer/sm-shop/pom.xml
@@ -205,6 +205,12 @@
             <version>${revision}</version>
         </dependency>
 
+        <!-- Apache ActiveMQ -->
+        <dependency>
+            <groupId>org.apache.activemq</groupId>
+            <artifactId>activemq-all</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
