+++ projects/stage-1/shopizer/sm-shop/src/main/java/com/salesmanager/shop/application/ShopApplication.java
@@ -1,5 +1,6 @@
 package com.salesmanager.shop.application;
 
+import org.apache.activemq.command.ActiveMQTextMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.ApplicationRunner;
@@ -11,6 +12,10 @@ import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
 
+import javax.jms.JMSException;
+import javax.jms.MessageNotWriteableException;
+import javax.jms.MessageProducer;
+import javax.jms.TextMessage;
 import javax.sql.DataSource;
 
 
@@ -33,6 +38,21 @@ public class ShopApplication extends SpringBootServletInitializer {
         return args -> {
             System.out.println("Get Connection : " + dataSource.getConnection());
         };
+    }
+
+    @Bean
+    @Autowired
+    public ApplicationRunner runner2(MessageProducer messageProducer) {
+        return args -> {
+            TextMessage textMessage = createTextMessage("Hello,World");
+            messageProducer.send(textMessage);
+        };
+    }
+
+    private TextMessage createTextMessage(String content) throws JMSException {
+        ActiveMQTextMessage textMessage = new ActiveMQTextMessage();
+        textMessage.setText(content);
+        return textMessage;
     }
 
 }
