+++ projects/stage-1/shopizer/sm-shop/src/main/java/com/salesmanager/shop/jms/MessageProducerFactory.java
@@ -1,0 +1,67 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.salesmanager.shop.jms;
+
+import javax.jms.*;
+import javax.naming.Context;
+import javax.naming.Name;
+import javax.naming.RefAddr;
+import javax.naming.Reference;
+import javax.naming.spi.ObjectFactory;
+import java.util.Hashtable;
+
+import static java.lang.String.valueOf;
+
+public class MessageProducerFactory implements ObjectFactory {
+
+    @Override
+    public Object getObjectInstance(Object obj, Name name, Context nameCtx,
+                                    Hashtable<?, ?> environment) throws Exception {
+
+        if (obj == null || !(obj instanceof Reference)) {
+            return null;
+        }
+
+        Reference reference = (Reference) obj;
+
+        String queueName = getAttribute(reference, "queueName");
+
+        String connectionFactoryJndiName = getAttribute(reference, "connectionFactoryJndiName");
+
+        // nameCtx 发现同级目录的 Context
+        ConnectionFactory connectionFactory = (ConnectionFactory) nameCtx.lookup(connectionFactoryJndiName);
+
+        Connection connection = connectionFactory.createConnection();
+        connection.start();
+
+        // Create a Session
+        Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
+
+        // Create the destination (Topic or Queue)
+        Destination destination = session.createQueue(queueName);
+
+        // Create a MessageProducer from the Session to the Topic or Queue
+        MessageProducer producer = session.createProducer(destination);
+
+        return producer;
+    }
+
+    private String getAttribute(Reference reference, String attributeName) {
+        RefAddr refAddr = reference.get(attributeName);
+        return refAddr == null ? null : valueOf(refAddr.getContent());
+    }
+}
